+++ src/main/java/wooteco/subway/auth/exception/LoginFailException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception.auth;
+package wooteco.subway.auth.exception;
 
 public class LoginFailException extends RuntimeException {
