+++ src/main/java/wooteco/subway/auth/exception/LoginWrongPasswordException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception.auth;
+package wooteco.subway.auth.exception;
 
 public class LoginWrongPasswordException extends LoginFailException {
