+++ src/main/java/wooteco/subway/member/application/MemberService.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.auth.dto.TokenRequest;
 import wooteco.subway.auth.dto.TokenResponse;
 import wooteco.subway.auth.infrastructure.JwtTokenProvider;
-import wooteco.subway.exception.auth.LoginFailEmailException;
-import wooteco.subway.exception.auth.LoginWrongPasswordException;
-import wooteco.subway.exception.member.NotRegisteredMemberException;
+import wooteco.subway.auth.exception.LoginFailEmailException;
+import wooteco.subway.auth.exception.LoginWrongPasswordException;
+import wooteco.subway.member.exception.NotRegisteredMemberException;
 import wooteco.subway.member.dao.MemberDao;
 import wooteco.subway.member.domain.Member;
 import wooteco.subway.member.dto.MemberRequest;
