+++ src/main/java/wooteco/subway/path/ui/PathControllerAdvice.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.exception.path.NotReachableException;
+import wooteco.subway.path.exception.NotReachableException;
 
 @RestControllerAdvice
 public class PathControllerAdvice {
